VALID COMMIT
We have finished the post-commit phase. It is a bit messy (to say the
least), but it appears to be sound.

We also fixed a couple of permission issues among subvolumes and
snapshots, and we are now able to execute tests without root as long
as the TxSv's permissions allow us to do so.